pyproject.toml
[tool.poetry]
name = "llm-merging"
version = "0.1.0"
description = ""
authors = ["Your Name <email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10.14"
absl-py="2.1.0"
accelerate="0.29.1"
aiohttp="3.9.3"
aiosignal="1.3.1"
altair="5.3.0"
annotated-types="0.6.0"
anyio="4.3.0"
appdirs="1.4.4"
astor="0.8.1"
asttokens="2.4.1"
async-timeout="4.0.3"
attrs="23.2.0"
base58="2.1.1"
black="21.12b0"
blinker="1.7.0"
brotli="1.1.0"
cachetools="5.3.3"
certifi="2024.2.2"
click="7.1.2"
datasets="2.18.0"
decorator="5.1.1"
dill="0.3.8"
distro="1.9.0"
evaluate="0.4.1"
exceptiongroup="1.2.0"
executing="2.0.1"
filelock="3.13.1"
fire="0.6.0"
flake8="7.0.0"
frozenlist="1.4.1"
fsspec="2024.2.0"
#git+https://github.com/bigscience-workshop/promptsource"
gitdb="4.0.11"
gitpython="3.1.43"
gmpy2="2.1.2"
h11="0.14.0"
httpcore="1.0.5"
httpx="0.27.0"
huggingface-hub="0.22.2"
idna="3.4"
inflate64="1.0.0"
iniconfig="2.0.0"
ipdb="0.13.13"
ipython="8.23.0"
isort="5.8.0"
jedi="0.19.1"
jinja2="3.1.3"
joblib="1.4.0"
jsonschema-specifications="2023.12.1"
jsonschema="4.21.1"
markupsafe="2.1.3"
matplotlib-inline="0.1.6"
mccabe="0.7.0"
mkl-fft="1.3.8"
mkl-random="1.2.4"
mkl-service="2.4.0"
mpmath="1.3.0"
multidict="6.0.5"
multipledispatch="1.0.0"
multiprocess="0.70.16"
multivolumefile="0.2.3"
mypy-extensions="1.0.0"
networkx="3.1"
nltk="3.8.1"
numpy="1.26.4"
openai="1.17.1"
packaging="24.0"
pandas="2.2.1"
parso="0.8.4"
pathspec="0.12.1"
peft="0.10.0"
pexpect="4.9.0"
pillow="10.2.0"
pip="24.0"
plotly="5.20.0"
pluggy="1.4.0"
prompt-toolkit="3.0.43"
protobuf="5.26.1"
psutil="5.9.8"
ptyprocess="0.7.0"
pure-eval="0.2.2"
py7zr="0.21.0"
pyarrow-hotfix="0.6"
pyarrow="15.0.2"
pybcj="1.0.2"
pycodestyle="2.11.1"
pycryptodomex="3.20.0"
pydantic-core="2.18.1"
pydantic="2.7.0"
pydeck="0.8.1b0"
pyflakes="3.2.0"
pygments="2.17.2"
pyppmd="1.1.0"
pytest="8.1.1"
python-dateutil="2.9.0.post0"
pytz="2024.1"
pyyaml="6.0.1"
pyzstd="0.15.10"
referencing="0.34.0"
regex="2023.12.25"
requests="2.31.0"
responses="0.18.0"
rouge_score="0.1.2"
rpds-py="0.18.0"
safetensors="0.4.2"
scikit-learn="1.4.2"
scipy="1.13.0"
setuptools="68.2.2"
six="1.16.0"
smmap="5.0.1"
sniffio="1.3.1"
stack-data="0.6.3"
streamlit="0.82.0"
sympy="1.12"
tempdir="0.7.1"
tenacity="8.2.3"
termcolor="2.4.0"
texttable="1.7.0"
threadpoolctl="3.4.0"
tokenizers="0.15.2"
toml="0.10.2"
tomli="1.2.3"
toolz="0.12.1"
torch="2.2.2"
torchaudio="2.2.2"
torchvision="0.17.2"
tornado="6.4"
tqdm="4.66.2"
traitlets="5.14.2"
transformers="4.39.3"
tree-sitter-languages="1.10.2"
tree-sitter="0.21.3"
triton="2.2.0"
typing-extensions="4.9.0"
tzdata="2024.1"
tzlocal="5.2"
urllib3="2.1.0"
validators="0.28.0"
watchdog="4.0.0"
wcwidth="0.2.13"
wget="3.2"
wheel="0.41.2"
xxhash="3.4.1"
yarl="1.9.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
